scripts/requirements.txt
fastapi
uvicorn
twilio
websockets
audioop-lts
